convert ghs over to use updated helpers etc
